Pagination optimizations
When compiling search results, only one item per OCN is needed (doesn't matter which orderItem it is, copy-level info will be attached to that item for any other orderItems). This de-duplication was moved to the MySQL queries rather than further down in the code, so as to prevent mis-pagination that is counting duplicated search results. Pagination links for format searches (instead of subject searches) were also fixed.