Improved date searches
dateLoaded is now used as an additional criteria when finding books with blank call numbers to reload/refresh.
When determining a cutoff date with "receipt" as the type of date used, large "monthPad" windows for a book to be in cataloging would result in incorrect dates being queried to MySQL. The code has been improved and generalized to handle situations where the cutoff date goes back one or more years from the present date.
When loading book results, receiptDate is now used as a search parameter to expedite load time (previously any book meeting non-date criteria would be loaded, and then date restrictions would only take effect when copy/LHR info was searched for each book).